requirements.txt
appnope==0.1.4
asttokens==3.0.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
fonttools==4.55.3
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
matplotlib==3.9.4
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
# Editable Git install with no remote (novel_toy_tools==1.0)
-e /Users/tendejan/Desktop/Tom Endejan Novel Toy 2024
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyOpenGL==3.1.7
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
scipy==1.13.1
six==1.17.0
stack-data==0.6.3
tdqm==0.0.1
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
zipp==3.21.0
